+++ diive/core/plotting/fitplot.py
@@ -8,6 +8,7 @@ def fitplot(
         edgecolor: str = 'none',
         marker: str = 'o',
         alpha: float = 1,
+        showfit:bool=True,
         show_prediction_interval:bool=True,
         size_scatter:int=60
 ):
@@ -24,7 +25,7 @@ def fitplot(
     line_xy = ax.scatter(flux_bts_results['x'],
                          flux_bts_results['y'],
                          edgecolor=edgecolor, color=color, alpha=alpha, s=size_scatter,
-                         label=label,
+                         label=_label,
                          zorder=1, marker=marker)
 
     # Highlight year
@@ -42,42 +43,42 @@ def fitplot(
                                     alpha=1, s=100,
                                     label=f"{label} {highlight_year} ({_numvals_y} days)",
                                     zorder=98, marker=marker)
+    else:
+        line_highlight = None
 
     # Fit
-    a = flux_bts_results['fit_params_opt'][0]
-    b = flux_bts_results['fit_params_opt'][1]
-    c = flux_bts_results['fit_params_opt'][2]
-    operator1 = "+" if b > 0 else ""
-    operator2 = "+" if c > 0 else ""
-    label_fit = rf"$y = {a:.4f}x^2{operator1}{b:.4f}x{operator2}{c:.4f}$"
-    line_fit, = ax.plot(flux_bts_results['fit_df']['fit_x'],
-                        flux_bts_results['fit_df']['nom'],
-                        c=color_fitline, lw=3, zorder=3, alpha=1, label=label_fit)
-
-    # Fit confidence region
-    # Uncertainty lines (95% confidence)
-    line_fit_ci = ax.fill_between(flux_bts_results['fit_df']['fit_x'],
-                                  flux_bts_results['fit_df']['nom_lower_ci95'],
-                                  flux_bts_results['fit_df']['nom_upper_ci95'],
-                                  alpha=.2, color=color_fitline, zorder=2,
-                                  label="95% confidence region")
-
-    # Fit prediction interval
+    line_fit = line_fit_ci = None
     line_fit_pb = None
-    if show_prediction_interval:
-        # Lower prediction band (95% confidence)
-        ax.plot(flux_bts_results['fit_df']['fit_x'],
-                flux_bts_results['fit_df']['lower_predband'],
-                color=color_fitline, ls='--', zorder=3, lw=2,
-                label="95% prediction interval")
-        # Upper prediction band (95% confidence)
-        line_fit_pb, = ax.plot(flux_bts_results['fit_df']['fit_x'],
-                               flux_bts_results['fit_df']['upper_predband'],
-                               color=color_fitline, ls='--', zorder=3, lw=2,
-                               label="95% prediction interval")
+    if showfit:
+        a = flux_bts_results['fit_params_opt'][0]
+        b = flux_bts_results['fit_params_opt'][1]
+        c = flux_bts_results['fit_params_opt'][2]
+        operator1 = "+" if b > 0 else ""
+        operator2 = "+" if c > 0 else ""
+        label_fit = rf"$y = {a:.4f}x^2{operator1}{b:.4f}x{operator2}{c:.4f}$"
+        line_fit, = ax.plot(flux_bts_results['fit_df']['fit_x'],
+                            flux_bts_results['fit_df']['nom'],
+                            c=color_fitline, lw=3, zorder=3, alpha=1, label=label_fit)
 
-    if highlight_year:
-        return line_xy, line_fit, line_fit_ci, line_fit_pb, line_highlight
+        # Fit confidence region
+        # Uncertainty lines (95% confidence)
+        line_fit_ci = ax.fill_between(flux_bts_results['fit_df']['fit_x'],
+                                      flux_bts_results['fit_df']['nom_lower_ci95'],
+                                      flux_bts_results['fit_df']['nom_upper_ci95'],
+                                      alpha=.2, color=color_fitline, zorder=2,
+                                      label="95% confidence region")
 
-    return line_xy, line_fit, line_fit_ci, line_fit_pb
+        # Fit prediction interval
+        if show_prediction_interval:
+            # Lower prediction band (95% confidence)
+            ax.plot(flux_bts_results['fit_df']['fit_x'],
+                    flux_bts_results['fit_df']['lower_predband'],
+                    color=color_fitline, ls='--', zorder=3, lw=2,
+                    label="95% prediction interval")
+            # Upper prediction band (95% confidence)
+            line_fit_pb, = ax.plot(flux_bts_results['fit_df']['fit_x'],
+                                   flux_bts_results['fit_df']['upper_predband'],
+                                   color=color_fitline, ls='--', zorder=3, lw=2,
+                                   label="95% prediction interval")
 
+    return line_xy, line_fit, line_fit_ci, line_fit_pb, line_highlight