+++ diive/pkgs/flux/__init__.py
@@ -1,2 +1,2 @@
-import diive.pkgs.flux.criticalheatdays
+import diive.pkgs.flux.criticaldays
 import diive.pkgs.flux.carboncost